TrackedBox.scala
